move starting logs in a new afterReady event (launched after all ready listeners)
