--- src/listeners/startLog.ready.native.ts
@@ -1,48 +1,0 @@
-// system file, please don't modify it
-
-import figlet from "figlet"
-import boxen from "boxen"
-import chalk from "chalk"
-
-import * as app from "../app.js"
-
-import { filename } from "dirname-filename-esm"
-
-const __filename = filename(import.meta)
-
-const listener: app.Listener<"ready"> = {
-  event: "ready",
-  description: "Just log that bot is ready",
-  once: true,
-  async run() {
-    setTimeout(() => {
-      app.log(
-        `ok i'm ready! ${chalk.blue(
-          "My default prefix is",
-        )} ${chalk.bgBlueBright.black(process.env.BOT_PREFIX)}`,
-      )
-
-      figlet(app.packageJSON.name, (err, value) => {
-        if (err) return app.error(err, __filename, true)
-
-        console.log(
-          boxen(chalk.blueBright(value), {
-            float: "center",
-            borderStyle: {
-              topLeft: " ",
-              topRight: " ",
-              bottomLeft: " ",
-              bottomRight: " ",
-              top: " ",
-              left: " ",
-              right: " ",
-              bottom: " ",
-            },
-          }),
-        )
-      })
-    }, 500)
-  },
-}
-
-export default listener
